relocate content-wrapper component
