+++ apps/app/src/routes/settings/user/user-detail.page.tsx
@@ -2,8 +2,8 @@ import { Space, Spin } from 'antd';
 import { useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import { PageHeader } from '../../../components/app/page-header.component';
+import ContentWrapper from '../../../components/content-wrapper.component';
 import { useLazyGetUserForSettingsQuery } from '../../../redux/services/graphql-api/settings/settings.api';
-import ContentWrapper from '../../home/components/content-wrapper.component';
 import { AssignRolesCard } from './components/assign-roles-card.component';
 import { OrgChartAccessCard } from './components/org-chart-access-card.component';
 import { OtherDetailsCard } from './components/other-details-card.component';
